Always update all views' internal state when state changes
it's common to do something like having an event that is attached inside a mount(), which at a later date will do something based on this.state - before this change, if a view wasn't showing, the state would be wrong, which is confusing. This change also means that you may use this.state inside showing, which previously was a source of a little confusion.